+++ src/components/users/UserController.js
@@ -4,7 +4,7 @@ import { Switch, Route } from "react-router-dom";
 import Login from "./Login";
 import Register from "./Register";
 import Dashboard from './Dashboard';
-import FriendProfile from "./friends-profile/FriendProfile";
+import FriendProfile from "./profile/FriendProfile";
 import PrivateRoute from "../../components/PrivateRoute";
