--- src/components/users/profile/userProfile.css
@@ -1,83 +1,0 @@
-.profile-image-ch {
-  border-radius: 50%;
-  width: 50%;
-  height: 50%;
-}
-
-.profile-img-name-ch {
-  display: flex;
-  align-items: center;
-  padding-bottom: 5%;
-  border-bottom: 1px solid grey;
-}
-
-.profile-name-ch {
-  font-size: 500%;
-  margin: 0;
-}
-.name-shop-ch {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-left: 20%;
-}
-
-.fav-shop-heading-ch {
-  color: grey;
-}
-
-.fav-shop-title-ch {
-  font-size: large;
-  min-width: 100%;
-  margin-top: 2%;
-}
-
-.user-bio-ch {
-  padding: 5%;
-  width: 100%;
-  height: 80px;
-  box-shadow: 0 2px 2px 0 rgba(0, 0, 0, 0.2);
-}
-
-.activity-container-ch {
-  margin-top: 5%;
-  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2);
-  width: 100%;
-  margin: 0;
-  overflow: scroll;
-  max-height: 1000px;
-}
-
-.bio-activity-ch {
-  display: flex;
-  margin-top: 5%;
-  width: 100%;
-}
-
-.act-card-ch {
-  width: 100%;
-}
-.head-act-ch {
-  text-align: center;
-  color: grey;
-}
-.picture-btn-ch {
-  display: flex;
-  flex-direction: column;
-}
-.profile-image-ch {
-  border-radius: 50%;
-  width: 100%;
-  height: 100%;
-}
-.img-ch-ch {
-  width: 7%;
-  height: 50px;
-  border-radius: 50%;
-  margin-right: 2%;
-}
-.img-head-ch-ch {
-  display: flex;
-  margin-bottom: 3%;
-  align-items: center;
-}
